requirements.txt
accelerate==0.26.0
autocommand==2.2.2
backports.tarfile==1.2.0
bitsandbytes==0.43.1
certifi==2025.1.31
charset-normalizer==3.4.1
cmake==3.31.4
filelock==3.17.0
fsspec==2025.2.0
fuzzywuzzy==0.18.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.5
json5==0.9.11
Levenshtein==0.26.1
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.1.1
pillow==10.2.0
platformdirs==4.2.2
psutil==6.1.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
six==1.17.0
sympy==1.13.3
tokenizers==0.21.0
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.65.0
transformers==4.48.1
triton==3.0.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.19.2
spacy
negspacy
spacy-langdetect
thefuzz
en-core-sci-lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_lg-0.5.1.tar.gz#sha256=b64746e84ca033bf4158fb811d93d50cfd8f75fead85c7261e4e6119bacded22
langchain
langgraph
langchain-huggingface
IPython
datasets
tiktoken==0.9.0
orjson
omegaconf
pydantic<2.0
